Fix some edge cases.

+++ src/deref.c
@@ -1,7 +1,6 @@
 #include "deref.h"
 #include "buffer.h"
 #include "trace.h"
-#include <errno.h>
 #include <stdlib.h>
 #include <unistd.h>
 
@@ -18,7 +17,7 @@ char *deref_fd(int fd, size_t length_guess, struct trace *trace) {
 
   for (;;) {
     char *target = TNULL(malloc(max_size), trace);
-    int length = TNEG(
+    ssize_t length = TNEG(
         readlink(get_string(get_view(link_buffer)), target, max_size), trace);
 
     if (!ok(trace)) {
